reorganized code, missed something
