+++ lib/response.js
@@ -1,17 +1,15 @@
 'use strict';
 
-const MessagingResponse = require('twilio').twiml.MessagingResponse;
 const messaging = module.exports = {};
 
 messaging.sendAndWrite = function(res, message){
-  let twiml = messaging.send(message);
+  let twiml = messaging.send(req, message);
   messaging.end(res, twiml);
 };
 
-messaging.send = function(message){
-  let twiml = new MessagingResponse();
-  twiml.message(message);
-  return (twiml);
+messaging.send = function(req, message){
+  req.twiml.message(message);
+  return (req.twiml);
 };
 
 messaging.end = function(res, twiml) {
