+++ routes/sms.js
@@ -5,7 +5,6 @@ const bodyParser = require('body-parser').urlencoded({
 });
 const smsRouter = module.exports = express.Router();
 
-const MessagingResponse = require('twilio').twiml.MessagingResponse;
 const messaging = require('../lib/response');
 
 const townHallHandler = require('./townHallMiddleware');
@@ -18,12 +17,11 @@ smsRouter.post('/sms',
   getEvents,
   (req, res) => {
     if (req.townHalls.length > 0) {
-      let twiml = new MessagingResponse();
       req.townHalls.forEach((townhall) => {
-        twiml.message(townhall.print());
+        req.twiml.message(townhall.print());
       });
-      twiml.message('That\'s all the events we have for your reps');
-      return messaging.end(res, twiml);
+      req.twiml.message('That\'s all the events we have for your reps');
+      return messaging.end(res, req.twiml);
     }
-    messaging.sendAndWrite(res, 'There are not any upcoming town halls in your area.');
+    messaging.sendAndWrite(req, res, 'There are not any upcoming town halls in your area.');
   });
